Determine the branches to be compared from the current repo
not the one about to be cloned

When running in CI systems we start with an existing local copy of the repo.
If we clone it we lose track of its origin branches.  Prior to this commit
we hacked around this problem by explicitly checking out all the origin
branches in the existing local copy.  This commit provides a nicer solution.
We use the originBranchesHashes_ parameter to pass in a function which looks
up the branches in the existing local copy.  It's fine to do this because
the cloned copy contains the relevant commits, its branches just don't refer
to those commits.